reset btn fixed after refactor
