+++ src/components/layout/Navbar.tsx
@@ -19,15 +19,24 @@ interface NavbarProps {
 const Navbar: FC<NavbarProps> = ({ className }) => {
   const [isMobileMenuOpen, setIsMobileMenuOpen] = useState(false);
   const [isDropdownOpen, setIsDropdownOpen] = useState(false);
-  const [userAvatar, setUserAvatar] = useState<string | null>(
-    typeof window !== "undefined" ? localStorage.getItem("userAvatar") : null
-  );
+  const [userAvatar, setUserAvatar] = useState<string | null>(null);
+  const [mounted, setMounted] = useState(false);
 
   useEffect(() => {
-    if (localStorage.getItem("userAvatar")) {
-      setUserAvatar(localStorage.getItem("userAvatar"));
+    setMounted(true);
+
+    if (typeof window !== "undefined") {
+      if (localStorage.getItem("userAvatar")) {
+        setUserAvatar(localStorage.getItem("userAvatar"));
+      }
     }
-  }, [userAvatar]);
+  }, []);
+
+  // useEffect(() => {
+  //   if (localStorage.getItem("userAvatar")) {
+  //     setUserAvatar(localStorage.getItem("userAvatar"));
+  //   }
+  // }, [userAvatar]);
 
   const dropdownRef = useRef<HTMLDivElement>(null);
 
@@ -131,23 +140,32 @@ const Navbar: FC<NavbarProps> = ({ className }) => {
               ref={dropdownRef}
               onMouseEnter={() => setIsDropdownOpen(true)}
             >
-              <button className="focus:outline-none flex items-center">
-                {userAvatar ? (
-                  <Image
-                    src={userAvatar}
-                    alt="User Avatar"
-                    width={60}
-                    height={60}
-                    className=" h-14 w-14 rounded-full border-2 border-neonGreen cursor-pointer"
-                  />
-                ) : (
-                  <Image
-                    src="/icons/avatar-default.jpg"
-                    alt="Default Avatar"
-                    width={60}
-                    height={60}
-                    className="rounded-full border-2 border-neonGreen cursor-pointer"
-                  />
+              <button className="focus:outline-none flex items-center h-[60px] w-[60px]">
+                {mounted &&
+                  (userAvatar ? (
+                    <Image
+                      src={userAvatar}
+                      alt="User Avatar"
+                      width={60}
+                      height={60}
+                      className="h-14 w-14 rounded-full cursor-pointer border-2 border-neonGreen"
+                    />
+                  ) : (
+                    <Image
+                      src="/icons/avatar-default.jpg"
+                      alt="Default Avatar"
+                      width={60}
+                      height={60}
+                      className="h-14 w-14 rounded-full cursor-pointer border-2 border-neonGreen"
+                    />
+                  ))}
+
+                {!mounted && (
+                  <div
+                    className={`rounded-full z-2 border-2 border-neonGreen cursor-pointer z-2 h-14 w-14 bg-green ${
+                      mounted ? "bg-opacity-0" : "bg-opacity-20 animate-pulse"
+                    }`}
+                  ></div>
                 )}
               </button>
               {isDropdownOpen && (
